+++ src/pages/SiteNewPage.tsx
@@ -3,7 +3,7 @@ type Props = {}
 export const SiteNew = ({}: Props) => {
   return (
     <div data-testid="siteNew-page">
-      <p>SiteNew</p>
+      <p>Site New</p>
     </div>
   )
 }
